+++ CHANGELOG.md
@@ -7,9 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [8.9.0] - 2025-08-11
+
 ### Added
 
 - arc gis map snapshot images
+
+### Dependencies
+
+- update [@oneblink/sdk-core](https://www.npmjs.com/package/@oneblink/sdk-core) to [8.7.0-beta.2](https://github.com/oneblink/sdk-core-js/blob/master/CHANGELOG.md) (from [8.6.0](https://github.com/oneblink/sdk-core-js/releases/tag/v8.6.0))
 
 ## [8.8.0] - 2025-08-04